requirements.txt
requests
python-dotenv
InquirerPy